Add Checkstyle checks
Add Checkstyle checks for code. The set of rules are copied exactly from
RobotPathfinder. All Javadoc checks are suppressed for all classes not
in the com.arctos6135.stdplug.api package as they will not be part of
the public API and does not need Javadoc documentation.

This resolves issue #2. The next commit will add the documentation and
fix all the checkstyle warnings.